[IL2Asm] Change the arguments to WriteMemory to be more consistent with the actual generated asm

+++ CPUHelper/CPU.cs
@@ -57,7 +57,7 @@ namespace CPUHelper
 
         }
 
-        public static void WriteMemory(int addr, int c)
+        public static void WriteMemory(int addr, ushort c)
         {
 
         }
